requirements/test.txt
# production and test requirements

-r prod.txt

pytest>=4.6 # required for pytest-cov
pytest-django
pytest-cov
wheel
mysqlclient # tested against in CI for now
django-dbml>=0.4.0
percy-python-selenium
